Move reusable constants to the tools/constants.js file
